+++ app/routes/components/TimeSelector/changelog.codegen.ts
@@ -1,3 +1,4 @@
 export const changelog = ''
   + '- 2025-09-11: Initial version.\n'
+  + '- 2025-09-25: Improved TimeSelector styles for better appearance.\n'
   ;