+++ src/components/TimeSelector/index.ts
@@ -1,2 +1,4 @@
+// TimeSelector from copy-ui @ 2025-09-25.
+
 export { TimeSelector } from './TimeSelector';
 export type { TimeRange } from './TimeSelector';