+++ src/components/TimeSelector/CHANGELOG.md
@@ -1,1 +1,2 @@
 - 2025-09-11: Initial version.
+- 2025-09-25: Improved TimeSelector styles for better appearance.